资产来源，减少方式，减少原因，技术状态模块完善

+++ erahub_manage_asset/src/main/java/com/erahub/asset/metadata/converter/AssetCategoryConverter.java
@@ -51,7 +51,7 @@ public class AssetCategoryConverter {
 
         BeanUtils.copyProperties(assetCategoryDTO, assetCategory);
 
-        assetCategory.setLevel(Long.valueOf(assetCategoryDTO.getAssetCategoryId().length() / 2));
+        assetCategory.setLevel(Integer.valueOf(assetCategoryDTO.getAssetCategoryId().length() / 2));
     }
 
     /**
@@ -67,7 +67,7 @@ public class AssetCategoryConverter {
                 AssetCategory assetCategory = new AssetCategory();
                 BeanUtils.copyProperties(assetCategoryDTO, assetCategory);
 
-                assetCategory.setLevel(Long.valueOf(assetCategoryDTO.getAssetCategoryId().length() / 2));
+                assetCategory.setLevel(Integer.valueOf(assetCategoryDTO.getAssetCategoryId().length() / 2));
 
                 assetCategories.add(assetCategory);
             }
