+++ erahub_manage_common/src/main/java/com/erahub/common/vo/asset/metadata/AssetCategoryVO.java
@@ -22,11 +22,11 @@ public class AssetCategoryVO {
 
     private Long assetQuantity;
 
-    private Long level;
+    private Integer level;
 
-    private Long detailed;
+    private Integer detailed;
 
-    private Long status;
+    private Integer status;
 
     private Long depreciationMethodId;
